requirements.txt
# Core dependencies
pyyaml>=6.0
numpy>=1.24

# Analysis & visualization
pandas>=2.0
matplotlib>=3.7
scipy>=1.10

# Model providers (install as needed)
openai>=1.0
anthropic>=0.40

# For open-source models (install as needed)
# transformers>=4.40
# torch>=2.0
# vllm>=0.4  # For faster inference
